+++ src/router.ts
@@ -16,10 +16,11 @@ import {
   isConnected,
   writeResponse,
 } from '@browserless.io/browserless';
+import { EventEmitter } from 'events';
 import micromatch from 'micromatch';
 import stream from 'stream';
 
-export class Router {
+export class Router extends EventEmitter {
   protected log = createLogger('router');
   protected verbose = createLogger('router:verbose');
   protected httpRoutes: Array<HTTPRoute | BrowserHTTPRoute> = [];
@@ -29,7 +30,9 @@ export class Router {
     protected config: Config,
     protected browserManager: BrowserManager,
     protected limiter: Limiter,
-  ) {}
+  ) {
+    super();
+  }
 
   protected getTimeout(req: Request) {
     const timer = req.parsed.searchParams.get('timeout');
@@ -207,13 +210,6 @@ export class Router {
     return route;
   }
 
-  public teardown() {
-    this.httpRoutes = [];
-    this.webSocketRoutes = [];
-
-    return this.browserManager.stop();
-  }
-
   public getStaticHandler() {
     return this.httpRoutes.find((route) =>
       route.path.includes(HTTPManagementRoutes.static),
@@ -253,4 +249,19 @@ export class Router {
       (r.path as Array<PathTypes>).some((p) => micromatch.isMatch(pathname, p)),
     );
   }
+
+  /**
+   * Implement any browserless-core-specific shutdown logic here.
+   * Calls the empty-SDK stop method for downstream implementations.
+   */
+  public shutdown = async () => {
+    this.httpRoutes = [];
+    this.webSocketRoutes = [];
+    await this.stop();
+  };
+
+  /**
+   * Left blank for downstream SDK modules to optionally implement.
+   */
+  public stop = () => {};
 }